+++ elfutils/libelf/elf32_getshdr.c
@@ -124,8 +124,9 @@ elfw2(LIBELFBITS,getshdr) (Elf_Scn *scn)
       else if (elf->fildes != -1)
 	{
 	  /* Read the header.  */
-	  if (pread (elf->fildes, elf->state.ELFW(elf,LIBELFBITS).shdr, size,
-		     elf->start_offset + ehdr->e_shoff) != size)
+	  if ((size_t) pread (elf->fildes,
+			      elf->state.ELFW(elf,LIBELFBITS).shdr, size,
+			      elf->start_offset + ehdr->e_shoff) != size)
 	    {
 	      /* Severe problems.  We cannot read the data.  */
 	      __libelf_seterrno (ELF_E_READ_ERROR);
